[poincare] Fix confusion between double and float!

+++ poincare/include/poincare/confidence_interval.h
@@ -24,7 +24,7 @@ private:
   Expression * shallowReduce(Context& context, AngleUnit angleUnit) override;
   /* Evaluation */
   Expression * privateEvaluate(Expression::SinglePrecision p, Context& context, Expression::AngleUnit angleUnit) const override { return templatedEvaluate<float>(context, angleUnit); }
-  Expression * privateEvaluate(Expression::DoublePrecision p, Context& context, Expression::AngleUnit angleUnit) const override { return templatedEvaluate<float>(context, angleUnit); }
+  Expression * privateEvaluate(Expression::DoublePrecision p, Context& context, Expression::AngleUnit angleUnit) const override { return templatedEvaluate<double>(context, angleUnit); }
   template<typename T> Expression * templatedEvaluate(Context& context, AngleUnit angleUnit) const;
 };
